\name{stanford.machine.learning-package}
\alias{stanford.machine.learning-package}
\alias{stanford.machine.learning}
\docType{package}
\title{
Provides code to accompany Stanford CS229 lectures on "Machine Learning".
}
\description{
Provides code to accompany Stanford CS229 lectures on "Machine Learning".
}
\details{
\tabular{ll}{
Package: \tab stanford.machine.learning\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-02\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
